# ASSIGNMENT_12/docs/openapi.yaml
openapi: 3.0.3
info:
  title: UniMatch API
  description: |
    REST API for the UniMatch university application matching system.
    
    Manages learner profiles, university programmes, and application workflows.
    
    ## Architecture
    This API follows a layered architecture:
    - **Entities** (Assignment 9) – Domain model (LearnerProfile, UniversityProgramme, Application, etc.)
    - **Repositories** (Assignment 11) – Persistence abstraction (InMemory implementations)
    - **Services** (Assignment 12) – Business logic enforcement
    - **API** (Assignment 12) – REST endpoints exposed via FastAPI
    
    ## Business Rules
    - A learner may have at most **5 active** applications at any time.
    - Duplicate applications to the same programme are rejected.
    - Applications can only be made to **Published** programmes with a future deadline.
    - APS is calculated using the NSC point scale (top 6 subjects).
  version: 1.0.0
  contact:
    name: UniMatch Team
servers:
  - url: http://localhost:8000
    description: Local development server
...
paths:
...
  /api/learners:
    get:
      tags: [Learners]
      summary: List all learner profiles
      operationId: listLearners
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LearnerResponse'

    post:
      tags: [Learners]
      summary: Create a learner profile
      operationId: createLearner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLearnerRequest'
            example:
              full_name: "Thabo Nkosi"
              school_id_number: "2024-0001"
              grade: 12
              counselor_id: "00000000-0000-0000-0000-000000000001"
              school_id: "00000000-0000-0000-0000-000000000002"
      responses:
        '201':
          description: Learner profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearnerResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/learners/{learner_id}:
    parameters:
      - $ref: '#/components/parameters/LearnerIdParam'

    get:
      tags: [Learners]
      summary: Get a learner profile
      operationId: getLearner
      responses:
        '200':
          description: Learner profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearnerResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Learners]
      summary: Update a learner profile
      operationId: updateLearner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLearnerRequest'
      responses:
        '200':
          description: Updated learner profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearnerResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Learners]
      summary: Delete a learner profile
      operationId: deleteLearner
      responses:
        '204':
          description: Learner deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /api/learners/{learner_id}/marks:
    parameters:
      - $ref: '#/components/parameters/LearnerIdParam'
    post:
      tags: [Learners]
      summary: Add a mark to a learner profile
      operationId: addMark
      description: |
        Adds a subject mark to the learner. The APS score is recalculated 
        automatically using the NSC 7-point scale.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMarkRequest'
            example:
              subject_name: "Mathematics"
              score: 75
              exam_type: "NSC"
              academic_year: 2024
      responses:
        '201':
          description: Mark added; returns updated learner profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearnerResponse'
        '400':
          description: Invalid mark score (must be 0–100)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/learners/{learner_id}/aps:
    parameters:
      - $ref: '#/components/parameters/LearnerIdParam'
    get:
      tags: [Learners]
      summary: Get a learner's APS score
      operationId: getAps
      description: |
        Calculates and returns the NSC APS score based on the learner's 
        recorded marks. Top 6 subjects are used when more than 6 marks exist.
      responses:
        '200':
          description: APS score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApsResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/learners/{learner_id}/deactivate:
    parameters:
      - $ref: '#/components/parameters/LearnerIdParam'
    post:
      tags: [Learners]
      summary: Deactivate a learner profile
      operationId: deactivateLearner
      description: |
        Deactivates the learner profile and automatically cancels all 
        non-terminal active applications.
      responses:
        '200':
          description: Deactivated learner profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearnerResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ──────────────────────────────────────────────────────────────────────────
  # Programmes
  # ──────────────────────────────────────────────────────────────────────────
  /api/programmes:
    get:
      tags: [Programmes]
      summary: List all programmes
      operationId: listProgrammes
      parameters:
        - name: published_only
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: If true, return only currently published, non-expired programmes
      responses:
        '200':
          description: List of programmes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProgrammeResponse'

    post:
      tags: [Programmes]
      summary: Create a university programme
      operationId: createProgramme
      description: Creates a new programme in **Draft** status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProgrammeRequest'
      responses:
        '201':
          description: Programme created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgrammeResponse'
  /api/programmes/{programme_id}:
    parameters:
      - $ref: '#/components/parameters/ProgrammeIdParam'

    get:
      tags: [Programmes]
      summary: Get a programme
      operationId: getProgramme
      responses:
        '200':
          description: Programme details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgrammeResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Programmes]
      summary: Update a programme
      operationId: updateProgramme
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProgrammeRequest'
      responses:
        '200':
          description: Updated programme
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgrammeResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Programmes]
      summary: Delete a programme
      operationId: deleteProgramme
      responses:
        '204':
          description: Programme deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /api/programmes/{programme_id}/publish:
    parameters:
      - $ref: '#/components/parameters/ProgrammeIdParam'
    post:
      tags: [Programmes]
      summary: Publish a programme
      operationId: publishProgramme
      description: |
        Publishes a Draft programme so learners can apply.
        **Fails with 400** if the application deadline has already passed.
      responses:
        '200':
          description: Published programme
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgrammeResponse'
        '400':
          description: Deadline has passed; cannot publish
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/programmes/{programme_id}/deactivate:
    parameters:
      - $ref: '#/components/parameters/ProgrammeIdParam'
    post:
      tags: [Programmes]
      summary: Deactivate a programme
      operationId: deactivateProgramme
      responses:
        '200':
          description: Deactivated programme
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgrammeResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/programmes/{programme_id}/extend-deadline:
    parameters:
      - $ref: '#/components/parameters/ProgrammeIdParam'
    post:
      tags: [Programmes]
      summary: Extend the application deadline
      operationId: extendDeadline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExtendDeadlineRequest'
      responses:
        '200':
          description: Programme with updated deadline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgrammeResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/programmes/{programme_id}/eligibility:
    parameters:
      - $ref: '#/components/parameters/ProgrammeIdParam'
    get:
      tags: [Programmes]
      summary: Check a learner's eligibility
      operationId: checkEligibility
      description: |
        Returns an eligibility category based on the learner's APS score vs 
        the programme's minimum APS requirement.
        
        | Condition            | Result       |
        |----------------------|--------------|
        | APS ≥ min + 4        | Guaranteed   |
        | APS ≥ min + 1        | Likely       |
        | APS = min            | Borderline   |
        | APS < min            | NotEligible  |
      parameters:
        - name: learner_aps
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
            maximum: 42
          description: The learner's APS score
      responses:
        '200':
          description: Eligibility result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EligibilityResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ──────────────────────────────────────────────────────────────────────────
  # Applications
  # ──────────────────────────────────────────────────────────────────────────
...
  /api/applications/{application_id}:
    parameters:
      - $ref: '#/components/parameters/ApplicationIdParam'

    get:
      tags: [Applications]
      summary: Get an application
      operationId: getApplication
      responses:
        '200':
          description: Application details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Applications]
      summary: Hard-delete an application
      operationId: deleteApplication
      description: Permanently deletes an application record. Prefer `/cancel` for soft cancellation.
      responses:
        '204':
          description: Application deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  parameters:
    LearnerIdParam:
      name: learner_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProgrammeIdParam:
      name: programme_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
  schemas:
...
    LearnerResponse:
      type: object
      properties:
        learner_id:
          type: string
          format: uuid
        full_name:
          type: string
        school_id_number:
          type: string
        grade:
          type: integer
        status:
          type: string
          enum: [Creating, Incomplete, Active, MarksRecorded, RecommendationsGenerated, Stale, GuidanceComplete, Inactive]
        counselor_id:
          type: string
          format: uuid
        school_id:
          type: string
          format: uuid
        aps_score:
          type: integer
          description: Current NSC APS score (top 6 subjects)
        mark_count:
          type: integer
...